requirements.txt
requests==2.28.1
python-telegram-bot==20.0a2
lyricsgenius==3.0.1
spotipy==2.21.0
PyGithub==1.55
youtube-search-python==1.6.6
yt-dlp==2022.9.1
openai==0.26.3